requirements.txt
Flask==3.1.1
blinker==1.9.0
boto3==1.38.40
botocore==1.38.40
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.41
greenlet==3.2.3
typing_extensions==4.14.0
Werkzeug==3.1.3
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
s3transfer==0.13.0
six==1.17.0
Mako==1.3.10
alembic==1.16.2
Flask-Migrate==4.1.0
Flask-Mail==0.10.0
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
requests==2.32.4
urllib3==2.5.0
psycopg2-binary==2.9.10
Flask-WTF==1.2.2
WTForms==3.2.1
email_validator==2.2.0
dnspython==2.7.0
# For testing
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pygments==2.19.2
pytest==8.4.1
pytest-flask==1.3.0
flask_login==0.6.3
coverage==7.9.1
pytest-cov==6.2.1
pytest-mock==3.14.1